checkpoint wait()

+++ Labs/lab2-c-basics/pointers/main.c
@@ -1,8 +1,9 @@
 #include <stdio.h>
+#include <stdint.h>
 
 int main(void){
-    int a = 5;
-    int *b = &a;
+    int8_t a = 5;
+    int8_t *b = &a;
 
     printf("a = %d\n", a);
     *b = 10;
